评论，帖子分页

+++ src/main/java/com/example/Entity/Post.java
@@ -2,6 +2,8 @@ package com.example.Entity;
 
 
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
 import lombok.Data;
 import org.apache.ibatis.annotations.Mapper;
 
@@ -10,14 +12,16 @@ import java.time.LocalDateTime;
 
 @Data
 public class Post {
-    private Integer id;
+   //@JsonSerialize(using = ToStringSerializer.class)
+    private Long id;
+
     private String userEmail;
     private String title;
     private String content;
-    private int type;
-    private int status;
-    private int tag;
+    private Integer type;
+    private Integer status;
+    private Integer tag;
     private LocalDateTime createTime;
-    private int commentCount;
+    private Integer commentCount;
     private double score;
 }
